% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet_balances.R
\name{load_mat}
\alias{load_mat}
\title{Cointel Wallet Balances}
\usage{
load_mat(infile)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This pings the API via R to query blockchain balances
}
