% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet_balances.R
\name{wallet_balance_check}
\alias{wallet_balance_check}
\title{Cointel Wallet Balances}
\usage{
wallet_balance_check(wallet_address)
}
\arguments{
\item{wallet_address}{Path to the input file}
}
\value{
API call for wallet balances
}
\description{
This pings the API via R to query blockchain balances
}
